Initial attempts at reformatting version index page.
Now show number of computers tested, number of checksums, number of passing computers, number of failing computers, and error types for failing computers. This is NOT READY FOR DEPLOYMENT. Only mixed test cases are looking "right", and current calls to VersionsController#index are extremely costly to the database; need to add columns to test_case_versions to speed things up dramatically.