Make speed improvements to commit view and add support for pull requests
Pull requests are updated by the api_update_pull_requests method of Commit. They are treated as normal commits, but commits now have two new boolean attributes, pull_request and open. The update process now queries github for open pull requests, closes any open pull requests in the database that are not already marked as closed, and then creates any open pull requests that weren't already in existence. Pull requests do not have parents or children, but they are associated with all branches that their parents are (parents determined from github, but not saved to db).

Now open pull requests appear at the top of the commit dropdown menu for a branch in the main commit view if they are present.