Interface and url cleanup for commit views.
Dates were overly verbose, so they are shortened now and display in a logged
in user's time zone.

Also properly pass on user names and commit times for the commit view's
commit selector dropdown.

Finally, generalize using "head" in commit view URLs so that it works for any
valid branch (or defaults to main if an invalid branch is used).