Purge two unused features: pull requests and commit relationships
Pull requests were represented in the database as the special "merge commits"
that GitHub kinda/sorta creates whenever a pull request is opened (and then
when any commit is made on the branch to be merged into). They were a nice way
to see if a pull request would cause any issues, but they were triggering far
too many automated tests and open a door for maclicious code to be
automatically executed by external pull requests. No one is testing them anymore, and they have been completely removed from the interface and GitHub webhooks infrastructure. Old merge commits are still in the database, and the "pull request" and "open" columns still exist in the commit database so they can still be sought out if need be.

Commit relations were used to indicate parent/child relationships between commits. This feature was never used in practice, so the whole table has been removed from the database, as well as the relevant model and association declarations on the commit model.