Further tweaks to commit dropdown and display of commit messages.
Now commit messages that are sufficiently long are truncated and made scrollable.

Also fixes an issue where a commit that was made long before it was pushed
could cause it to appear like the head commit in the local commit (reaching
out to GitHub api for a time window around the commit time of the commit
could miss the commit itself). Getting this right sometimes results in a
longer hang while it makes multiple queries to the GitHub api to get a more
expansive list of commits, but it will now be much more likely to actually
get the right nearby commits.