Edit test case commit view to show inlist-level details.
Each inlist now has its own table that is toggled by navigation pills.
This doesn't appear if there is only a single inlist. The checkboxes
apply to all tables. Note that some are pointless at the inlist level
(RAM, checksums are not implemented at the inlist level). An unrleated
pre-existing problem is that the overview-level (test instance level) of
certain values like retries and steps are ambiguous, as different
multi-inlist test cases treat these quantities differently (yielding
per-inlist values vs. cumulative values). Not a quick fix for this, and
the interface should probably be tweaked to indicate this.