More tweaking of testhub/github interface.
Prefer ordering of commits in commits index by creation timestamp, and then
commit time as tie-breaker. This may cause innacurate orders for commits that
were ingested from github with a finite time between them (order is not
guaranteed). The new ingestion scheme ensures all commits have the same
creation time, so this should work better going forward.

Still not certain this is the way to go. Awaiting divergence from the
production app to see if this works reliably.